feat: finished porting tests and added script to run them all
