+++ platformio.ini
@@ -150,3 +150,21 @@ extends = env:tests
 build_flags =
     ${env:tests.build_flags}
     -DIDENTITY_TESTS=1
+
+[env:led_tests]
+extends = env:tests
+build_flags =
+    ${env:tests.build_flags}
+    -DLED_TESTS=1
+
+[env:programmer_tests]
+extends = env:tests
+build_flags =
+    ${env:tests.build_flags}
+    -DPROGRAMMER_TESTS=1
+
+[env:volt_tests]
+extends = env:tests
+build_flags =
+    ${env:tests.build_flags}
+    -DVOLT_TESTS=1
