--- tests/conf.yaml
@@ -1,7 +1,0 @@
-:cmock:
-    :attributes:
-      - __funky
-      - __intrinsic
-    :when_ptr: :compare
-    :plugins: 
-      - :ignore
